+++ backend/crates/entity/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod annotation;
 pub mod collaborator;
+pub mod collaborator_disc_result;
 pub mod expectation_contract_item;
 pub mod feedback;
 pub mod feedback_behavior;